+++ tests/testthat/test-model_evaluation.R
@@ -42,7 +42,7 @@ test_that("choice prediction works", {
     alpha = 1:5, Sigma = 1
   )
   data <- train_test(data, test_proportion = 0.3)
-  model <- mcmc(data$train, R = 1000, seed = 1)
+  model <- fit_model(data$train, R = 1000, seed = 1)
   expect_snapshot(predict(model, overview = TRUE))
   expect_snapshot(predict(model, overview = FALSE))
   expect_snapshot(predict(model, data = data$test, overview = TRUE))
@@ -59,7 +59,7 @@ test_that("preference classification works", {
     seed = 1,
     C = 2
   )
-  model <- mcmc(data,
+  model <- fit_model(data,
     R = 1000,
     seed = 1,
     latent_classes = list("C" = 2)